requirements.txt
typer[all]>=0.12.0
httpx==0.26.0
pandas==2.2.0
rich==13.7.0
pytest==8.0.0
pytest==8.0.0
plotly==5.18.0
pyjwt==2.8.0
beautifulsoup4==4.12.3
reportlab==4.0.9
regex==2023.12.25
PySide6
PySide6-WebEngine
qasync==0.27.1
openpyxl==3.1.2
pygments==2.17.2
